+++ src/components/header/header.component.ts
@@ -172,6 +172,10 @@ export class DataTableHeaderComponent {
   calcNewSorts(column: any, prevValue: number, newValue: number): any[] {
     let idx = 0;
 
+    if(!this.sorts) {
+      this.sorts = [];
+    }
+
     const sorts = this.sorts.map((s, i) => {
       s = { ...s };
       if (s.prop === column.prop) idx = i;